Require the path to the zipfile for build_lambdas.sh

+++ scripts/deploy/build_lambdas.sh
@@ -2,16 +2,25 @@ echo "Creating AWS Lambda Deployment Zip"
 
 BASEDIR=$(readlink -f "$(dirname $0)/../..")
 
-if [ ! -d "$BASEDIR/deploy" ]; then
+if [[ ! -d $BASEDIR/deploy ]]; then
     mkdir "$BASEDIR/deploy"
 fi
 
-ZIPFILE="$BASEDIR/deploy/hsreplay.zip"
+ZIPFILE="$1"
+if [[ -z $ZIPFILE ]]; then
+	>&2 echo "Usage: $0 <ZIPFILE>"
+	exit 1
+fi
+ZIPFILE=$(readlink -f "$ZIPFILE")
+
 SITE_PACKAGES=$(python -c "import sys; print(sys.path[-1])")
+echo "site-packages is $SITE_PACKAGES"
+echo "Archiving to $ZIPFILE"
 
 rm -f "$ZIPFILE"
 
 git -C "$BASEDIR" archive --format=zip HEAD -o "$ZIPFILE"
+
 
 cd "$SITE_PACKAGES"
 zip -r "$ZIPFILE" ./*
